+++ Editor/Dialogue/BalloonInspector.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.Linq;
 using Sirenix.OdinInspector;
 using Sirenix.OdinInspector.Editor;
 using UnityEditor;
@@ -9,34 +10,47 @@ using UnityEngine.Timeline;
 
 namespace OcDialogue.Editor
 {
-    [CustomEditor(typeof(Balloon))]
+    [CustomEditor(typeof(Balloon))][CanEditMultipleObjects]
     public class BalloonInspector : OdinEditor
     {
-        DialogueNode _node;
-        Balloon _target;
+        Balloon[] _targets;
+        
+        DialogueNode[] _nodes;
+
         protected override void OnEnable()
         {
             base.OnEnable();
-            _target = target as Balloon;
+            _targets = targets.Cast<Balloon>().ToArray();
+            FindNode();
         }
 
         public override void OnInspectorGUI()
         {
             base.OnInspectorGUI();
-            if (_node == null) _node = FindNode();
-            if (_node != null)
+            if (_nodes != null)
             {
-                // TextField는 노드 내에서 바인딩으로 연결함.
-                _node.RefreshTitle();
-                _node.RefreshIcons();
+                for (int i = 0; i < _nodes.Length; i++)
+                {
+                    var node = _nodes[i];
+                    node.RefreshTitle();
+                    node.RefreshSubtitleLabel();
+                    node.RefreshDescription();
+                    node.RefreshIcons();    
+                }
+                
             }
         }
 
-        DialogueNode FindNode()
+        void FindNode()
         {
-            if(DialogueEditorWindow.Instance == null) return null;
-            if(DialogueEditorWindow.Instance.GraphView == null) return null;
-            return DialogueEditorWindow.Instance.GraphView.Nodes.Find(x => x.Balloon.GUID == _target.GUID);
+            if(DialogueEditorWindow.Instance == null) return;
+            if(DialogueEditorWindow.Instance.GraphView == null) return;
+            
+            _nodes = new DialogueNode[targets.Length];
+            for (int i = 0; i < _nodes.Length; i++)
+            {
+                _nodes[i] = DialogueEditorWindow.Instance.GraphView.Nodes.Find(x => x.Balloon.GUID == _targets[i].GUID);  
+            }
         }
     }
 }